feat [cache_manager, resource_manager, daemon]: set_weight now supports float rather than int

+++ bindings/c/cache_manager_ffi.cpp
@@ -50,7 +50,7 @@ int midas_pool_update_limit(cache_pool_t pool, uint64_t limit_in_bytes) {
   return 0;
 }
 
-int midas_pool_set_weight(cache_pool_t pool, int32_t weight) {
+int midas_pool_set_weight(cache_pool_t pool, float weight) {
   auto pool_ = reinterpret_cast<midas::CachePool *>(pool);
   if (!pool_)
     return -1;
