+++ daemon/inc/daemon_types.hpp
@@ -53,7 +53,7 @@ public:
   bool overcommit_region();
   bool free_region(int64_t region_id);
   bool update_limit(uint64_t new_limit);
-  void set_weight(int32_t weight);
+  void set_weight(float weight);
   bool force_reclaim(uint64_t new_limit);
   bool profile_stats();
 
@@ -73,7 +73,7 @@ private:
   Daemon *daemon_;
   uint64_t region_cnt_;
   uint64_t region_limit_;
-  int32_t weight_;
+  float weight_;
   int32_t warmup_ttl_;
 
   friend class Daemon;
